permission edit done

+++ app/Http/Controllers/DepartmentController.php
@@ -3,17 +3,19 @@
 namespace App\Http\Controllers;
 
 use App\Department;
+use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class DepartmentController extends Controller
 {
     public function index(){
+        $this->authorize('view', Department::class);
         $departments = Department::all();
         return view('department.index', compact('departments'));
-
     }
     public function create(){
+        $this->authorize('create', Department::class);
         return view('department.create');
     }
     public function store(Request $request){
@@ -37,6 +39,7 @@ class DepartmentController extends Controller
     }
 
     public function edit($id){
+        $this->authorize('edit', Department::class);
         $department = Department::findOrfail($id);
         return view('department.edit', compact('department'));
     }
@@ -55,6 +58,7 @@ class DepartmentController extends Controller
         return back()->with($notification);
     }
     public function delete($id){
+        $this->authorize('delete', Department::class);
         $department = Department::findOrfail($id);
         $department->delete();
         $notification = array(
